test: split embedding lane tests
